refactor(web): simplify create release component boundaries
